refactor: reorganize with atomic design
