+++ apps/client/tsconfig.json
@@ -63,7 +63,7 @@
       "@home-assets/*": [
         "src/app/pages/home/assets/*"
       ],
-      "@home-components/*":[
+      "@home-components/*": [
         "src/app/pages/home/components/*"
       ],
     }
